requirement.txt
pandas
numpy
matplotlib
scikit-learn
keras
tensorflow
yfinance
Pillow
requests
streamlit